Move final Character files to the engine. Refactoring is complete.
