+++ ControlDataIntegrityTests/RelatedCharacterDataTests.cs
@@ -1,5 +1,4 @@
 ﻿using System.Linq;
-using Glyssen.Character;
 using GlyssenEngine.Character;
 using NUnit.Framework;
