+++ DevTools/CharacterDetailProcessing.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Text.RegularExpressions;
-using Glyssen.Character;
 using Glyssen.RefTextDevUtilities;
 using GlyssenEngine.Character;
 using SIL.Scripture;
