--- Glyssen/Character/ICharacterGroupSource.cs
@@ -1,7 +1,0 @@
-﻿namespace Glyssen.Character
-{
-	public interface ICharacterGroupSource
-	{
-		CharacterGroupTemplate GetTemplate(int numberOfActors);
-	}
-}
