+++ GlyssenEngine/Character/Template/ICharacterGroupSource.cs
@@ -1,0 +1,12 @@
+﻿/// <summary>
+///  This code is not compiled.
+///  Experimental code. Not used at this time.
+/// </summary>
+
+namespace GlyssenEngine.Character.Template
+{
+	public interface ICharacterGroupSource
+	{
+		CharacterGroupTemplate GetTemplate(int numberOfActors);
+	}
+}
