+++ GlyssenEngine/GlyssenEngine.csproj
@@ -22,6 +22,12 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <Compile Remove="Character\Template\CharacterGroupTemplate.cs" />
+    <Compile Remove="Character\Template\CharacterGroupTemplateExcelFile.cs" />
+    <Compile Remove="Character\Template\ICharacterGroupSource.cs" />
+  </ItemGroup>
+
+  <ItemGroup>
     <None Remove="Resources\Annotations.txt" />
     <None Remove="Resources\BiblicalAuthors.txt" />
     <None Remove="Resources\BlockSplit.css" />
@@ -93,6 +99,12 @@
     <Content Include="Resources\RelatedCharacters.xml" />
     <Content Include="Resources\SampleMRK.xml" />
     <Content Include="Resources\Silent.wav" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <None Include="Character\Template\CharacterGroupTemplate.cs" />
+    <None Include="Character\Template\CharacterGroupTemplateExcelFile.cs" />
+    <None Include="Character\Template\ICharacterGroupSource.cs" />
   </ItemGroup>
 
   <ItemGroup>
