+++ GlyssenEngineTests/Properties/Resources.resx
@@ -175,6 +175,9 @@
   <data name="TestCharacterDetailOct2015" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>..\Resources\TestCharacterDetailOct2015.txt;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;Windows-1252</value>
   </data>
+  <data name="TestCharacterGroups" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\TestCharacterGroups.xlsx;System.Byte[], mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
+  </data>
   <data name="TestCharacterVerse" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>..\Resources\TestCharacterVerse.txt;System.String, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089;Windows-1252</value>
   </data>
