+++ GlyssenEngineTests/ViewModelTests/VoiceActorAssignmentViewModelTests.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Data;
 using System.Linq;
 using GlyssenEngine;
 using GlyssenEngine.Casting;
