+++ GlyssenSharedTests/BlockElementTests.cs
@@ -1,5 +1,4 @@
-﻿using Glyssen;
-using Glyssen.Shared;
+﻿using Glyssen.Shared;
 using NUnit.Framework;
 using SIL.TestUtilities;
 using SIL.Xml;
